requirements.txt
autopep8==1.5.7
click==8.0.0
Flask==2.0.0
Flask-WTF==0.14.3
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.17.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.0
WTForms==2.3.3
